Code refactoring
Bean bug fix,
Controller implementation &
few code enhancement in getRandomString
